Overdue commit
Schema: made department_name unique, made role title unique, department on role is now department_id, and role on employees is now varchar, referencing title.
seeds: department is now department_id, changed role id's to role titles,
server: Added a cancel option to all delete prompts, changed department to department_id for role creation, added some commas, changed roleID to roleTitle on employee creation, skipping the manager prompt acknowledges user as a manager now, users now are prevented from skipping name entry, and attempted to get the delete employee working. Currently getting a "column undefined" error when attempting to delete employee.

+++ src/server.ts
@@ -12,6 +12,7 @@ const app = express();
 app.use(express.urlencoded({extended: false}));
 app.use(express.json());
 
+// ------------------------------------------------------------------------------------------------
 // Using inquirer
 
 // Nested inquirer calls for INSERTing into departmentsqueries
@@ -29,6 +30,7 @@ function addDepartment(): void {
                 }else if(result) {
                     console.log("Data successfully entered!");
                 }
+                // allDepartmentNames.push(response.addDep);
                 allDepartmentNames.push(response.addDep);
                 startCli();
             })
@@ -81,7 +83,8 @@ for(let i = 0; i < queryArray.length; i++){
     allDepartmentNames[i] = queryArray[i].department_name;
     allDepartmentIDs[i] = queryArray[i].department_id;
 };
-console.log(`All departments: ${allDepartmentNames}`);
+allDepartmentNames.unshift("cancel");
+console.log(`All departments: ${allDepartmentNames} (I would like a cancel feature)`);
 
 // let depts = await pool.query('SELECT department_name FROM departments');
 // Storing ROLES in an array
@@ -93,17 +96,21 @@ for(let i = 0; i < querryArrayR.length; i++){
     allRoles[i] = querryArrayR[i].title;
     roleIDs[i] = querryArrayR[i].id;
 };
+allRoles.unshift("cancel");
 console.log(`All roles: ${allRoles}, and their IDs: ${roleIDs}`);
+
 // Storing ALL employees in an array
 let resultE = await pool.query('SELECT * FROM employees');
 let querryArrayE = resultE.rows;
 let allEmployees: string[] = [];
 let employeeIDs: number[] = [];
+let IDtoDelete: number; //using in the deleteEmployee method
 for(let i = 0; i < querryArrayE.length; i++){
     allEmployees[i] = querryArrayE[i].first_name + ", " + querryArrayE[i].last_name;
     employeeIDs[i] = querryArrayE[i].id;
-
 };
+allEmployees.unshift("cancel");
+
 // Storing MANAGERS in an array
 let resultM = await pool.query('SELECT first_name, last_name FROM employees WHERE manager_id=NULL');
 let queryArrayM = resultM.rows;
@@ -121,12 +128,12 @@ function addRole(): void {
             type: "input",
             name: "roleTitle",
             message: "What role would you like to add?",
-        }
+        },
         {
             type: "input",
             name: "roleSalary",
             message: "What is the role's salary?",
-        }
+        },
         {
             // type: "input",
             // name: "roleDepartment",
@@ -142,9 +149,9 @@ function addRole(): void {
         /* Neither id nor department_id work. ${resonse.roleDepartment} gives an input syntax error in the console. It's expecting an integer, but it
            recieved a string. The thing is, it needs an array to work. If presented with IDs and names, it throws an error.
            */
-          let place = allDepartmentNames.indexOf(response.roleDepartment);
+          let place = allDepartmentNames.indexOf(response.roleDepartment) +1;
         console.log(`${response.roleDepartment} is the role department.`);
-        pool.query(`INSERT INTO roles(title, salary, department) VALUES ('${response.roleTitle}', ${response.roleSalary}, '${place}')`,
+        pool.query(`INSERT INTO roles(title, salary, department_id) VALUES ('${response.roleTitle}', ${response.roleSalary}, '${place}')`,
             (err: Error, result: QueryResult) => {
                 if(err){
                     console.log(err);
@@ -156,6 +163,8 @@ function addRole(): void {
         });
     })
 };
+//----------------------------------------------------------------------------------------------------------
+// Add employee 
 // Nested inquirer calls for INSERTing into employees queries
 function addEmployee(): void {
     inquirer.prompt([
@@ -163,12 +172,12 @@ function addEmployee(): void {
             type: "input",
             name: "firstName",
             message: "Please enter the employee's first name.",
-        }
+        },
         {
             type: "input",
             name: "lastName",
             message: "Please enter the employee's last name.",
-        }
+        },
         // {
         //     type: "input",
         //     name: "roleID",
@@ -176,26 +185,40 @@ function addEmployee(): void {
         // }
         {
             type: "select",
-            name: "roleID",
-            message: "Please enter the employee's role ID.",
+            name: "roleTitle",
+            message: "Please enter the employee's role title.",
             choices: allRoles,
-        }
+        },
         {
             type: "input",
             name: "managerID",
-            message: "Please enter the employee's manager's ID.",
+            message: "Please enter the employee's manager's ID. If you are the manager, then press ENTER.",
         }
     ])
     .then((response) => {
-        pool.query(`INSERT INTO employees (first_name, last_name, role_id, manager_id) VALUES ('${response.firstName}', '${response.lastName}', ${response.roleID}, ${response.managerID})`, 
-            (err: Error, result: QueryResult) => {
-                if(err){
-                    console.log(err);
-                }else if(result) {
-                    console.log("Data successfully entered.");
-                }
-                startCli();
-            });
+        // let managersIndex = ;
+        // let chosenManager = employeeIDs[managersIndex];
+        // Preventing empty strings from being assigned to firstName and lastName
+        if(response.firstName === ''|| response.lastName === ''){
+            console.log("First name and last name cannot be blank. Please try again.");
+            startCli();
+        }else{
+            /* Leaving a prompt blank does not result in NULL, but in an empty string;
+               since managerID wants INT, it throws an error. */
+            if(response.managerID === ''){response.managerID = null};
+            pool.query(`INSERT INTO employees (first_name, last_name, role_title, manager_id) VALUES ('${response.firstName}', '${response.lastName}', '${response.roleTitle}', ${response.managerID})`, 
+                (err: Error, result: QueryResult) => {
+                    if(err){
+                        console.log(err);
+                        console.log(`The manager id: ${response.managerID}, type: ${typeof response.managerID}`);
+                    }else if(result) {
+                        allEmployees.push(response.firstName + ', ' + response.lastName);
+                        
+                        console.log("Data successfully entered.");
+                    }
+                    startCli();
+                });
+            };
         });
     };
 // Nested inquirer calls for UPDATE-ing employees queries
@@ -217,17 +240,21 @@ function deleteDepartment(): void {
         console.log(`result.departmentDeletion.id? ${response.departmentDeletion.id}`);
         /* response.departmentDeletion.id returns "undefined". Keep in mind that allDepartments is a string[]
            containing only the department names and not their IDs.*/
-        pool.query(`DELETE FROM departments WHERE department_name='${response.departmentDeletion}'`, (err: Error, result: QueryResult) => {
-            if(err){
-                console.log(err);
-            }else if(result){
-                console.log("Record successfully deleted.");
-            };
-            let deletedItem = allDepartmentNames.indexOf(response.departmentDeletion);
-            allDepartmentNames.splice(deletedItem, 1);
-            console.log(`allDpeartmentNames: ${allDepartmentNames}`);
+        if(response.departmentDeletion === "cancel"){
             startCli();
-        });
+        }else{
+            pool.query(`DELETE FROM departments WHERE department_name='${response.departmentDeletion}'`, (err: Error, result: QueryResult) => {
+                if(err){
+                    console.log(err);
+                }else if(result){
+                    console.log("Record successfully deleted.");
+                };
+                let deletedItem = allDepartmentNames.indexOf(response.departmentDeletion);
+                allDepartmentNames.splice(deletedItem, 1);
+                console.log(`allDpeartmentNames: ${allDepartmentNames}`);
+                startCli();
+            });
+        }
     })
 };
 
@@ -240,17 +267,21 @@ function deleteRole(): void {
             choices: allRoles
         }
     ]).then((response) => {
-        pool.query(`DELETE FROM roles WHERE title='${response.roleDeletion}'`, (err: Error, result: QueryResult) => {
-            if(err){
-                console.log(err);
-            }else if(result){
-                console.log("Record successfully deleted.");
-            }; 
-            let deletedItem = allRoles.indexOf(response.roleDeletion);
-            allRoles.splice(deletedItem, 1);
-            console.log(`allRoles: ${allRoles}`);
+        if(response.roleDeletion === "cancel"){
             startCli();
-        });
+        }else{
+            pool.query(`DELETE FROM roles WHERE title='${response.roleDeletion}'`, (err: Error, result: QueryResult) => {
+                if(err){
+                    console.log(err);
+                }else if(result){
+                    console.log("Record successfully deleted.");
+                }; 
+                let deletedItem = allRoles.indexOf(response.roleDeletion);
+                allRoles.splice(deletedItem, 1);
+                console.log(`allRoles: ${allRoles}`);
+                startCli();
+            });
+        }
     })
 };
 function deleteEmployee(): void {
@@ -262,14 +293,36 @@ function deleteEmployee(): void {
             choices: allEmployees
         }
     ]).then((response) => {
-        pool.query(`DELETE FROM employees WHERE id=${response}`, (err: Error, result: QueryResult) => {
-            if(err){
-                console.log(err);
-            }else if(result){
-                console.log("Record successfully deleted.");
-            };
+        if(response.employeeDeletion === "cancel"){
             startCli();
-        });
+        }else{
+            // I want to pull from employeeIds where the id matches the selected name
+            console.log(`The response: ${response}, ${typeof response}`); //type object
+            // time tto cookkkkk
+            pool.query(`SELECT * FROM employees`, (err: Error, result: QueryResult) => {
+                if(err){
+                    console.log(err)
+                }else if(result){
+                    let currentEmployees = result.rows;
+                    console.log(`resultRow length: ${currentEmployees.length}`); //it's 12??? of course it is
+                    let recordToDelete = currentEmployees.findIndex(record => `${record.first_name}, ${record.last_name}` === response.employeeDeletion);
+                    console.log(`The index of the person to terminate: ${recordToDelete}`);
+                    console.log(`The first_name of the 12th employee: ${currentEmployees[11].first_name}`);
+                    IDtoDelete = currentEmployees[recordToDelete].id;
+                    console.log(`The value of IDtoDelete: ${IDtoDelete}`);
+                }
+            })
+            // yea boiii
+            console.log(`Right before the DELETE query, this is IDtoDelete: ${IDtoDelete}`);
+            pool.query(`DELETE FROM employees WHERE id=${IDtoDelete}`, (err: Error, result: QueryResult) => {
+                if(err){
+                    console.log(err);
+                }else if(result){
+                    console.log("Record successfully deleted.");
+                };
+                startCli();
+            });
+        }
     })
 };
 // DELETE queries end
